Move import test to test_downstream
